Move SecondFactorTypeTranslationService to Stepup-bundle
As SelfService and RA will both use this service is will be moved to the
 stepup bundle.
